% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms.R
\name{archstabms}
\alias{archstabms}
\title{archstabms}
\usage{
archstabms(
  startStage = 1,
  stopStage = 6,
  base_dir = "/users/project/prj004631/afaure/DMS/Results/archstabms_proj",
  output_dir = ""
)
}
\arguments{
\item{startStage}{Start at a specified analysis stage (default:1)}

\item{stopStage}{Stop at a specified analysis stage (default:15)}

\item{base_dir}{Base directory for all input files (default:NB private CRG server path; change accordingly)}

\item{output_dir}{Output directory for all output files (default:same as base_dir)}
}
\value{
Nothing
}
\description{
Main analysis script.
}
