% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__create_dir.R
\name{archstabms__create_dir}
\alias{archstabms__create_dir}
\title{archstabms__create_dir}
\usage{
archstabms__create_dir(
  archstabms_dir,
  execute = TRUE,
  message = NULL,
  overwrite_dir = TRUE
)
}
\arguments{
\item{archstabms_dir}{directory path string (required)}

\item{execute}{whether or not the system command will be executed (required)}

\item{message}{message string (optional, default: NULL i.e. no message displayed)}

\item{overwrite_dir}{delete directory if already exists (optional, default: TRUE)}
}
\value{
Nothing
}
\description{
Create results folder for analysis script plots and saved objects.
}
