% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__define_confident_free_energies.R
\name{archstabms__define_confident_free_energies}
\alias{archstabms__define_confident_free_energies}
\title{archstabms__define_confident_free_energies}
\usage{
archstabms__define_confident_free_energies(
  input_dt,
  report_outpath,
  highlight_colour = "red",
  energy_max_ci = 1
)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{highlight_colour}{colour for highlights (default:red)}

\item{energy_max_ci}{maximum energy confidence interval (default:1)}
}
\value{
data.table with columns indicating confident free energies
}
\description{
Call confident ddGs.
}
