% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__fraction_bound.R
\name{archstabms__fraction_bound}
\alias{archstabms__fraction_bound}
\title{archstabms__fraction_bound}
\usage{
archstabms__fraction_bound(
  folding_energy,
  binding_energy,
  RT = 0.001987 * (273 + 24)
)
}
\arguments{
\item{folding_energy}{folding energies for 3-state model (required)}

\item{binding_energy}{binding energies for 3-state model (required)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
Fraction bound
}
\description{
Calculate fraction bound in 3-state model.
}
