% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__get_ambiguity_code.R
\name{archstabms__get_ambiguity_code}
\alias{archstabms__get_ambiguity_code}
\title{archstabms__get_ambiguity_code}
\usage{
archstabms__get_ambiguity_code(input_nucs)
}
\arguments{
\item{input_nucs}{vector of nucleotides (required)}
}
\value{
ambiguity code
}
\description{
Get IUPAC ambiguity codes.
}
