% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__get_codons.R
\name{archstabms__get_codons}
\alias{archstabms__get_codons}
\title{archstabms__get_codons}
\usage{
archstabms__get_codons(nt_seq)
}
\arguments{
\item{nt_seq}{nucleotide sequence string (required)}
}
\value{
Vector of codons
}
\description{
Split nucleotide sequence into codons.
}
