% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__get_model_results.R
\name{archstabms__get_model_results}
\alias{archstabms__get_model_results}
\title{archstabms__get_model_results}
\usage{
archstabms__get_model_results(input_folder, RT = 0.001987 * (273 + 24))
}
\arguments{
\item{input_folder}{path to input folder with model results (required)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
data.table with model results
}
\description{
Get model results and coefficients.
}
