% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__optimal_library_design.R
\name{archstabms__optimal_library_design}
\alias{archstabms__optimal_library_design}
\title{archstabms__optimal_library_design}
\usage{
archstabms__optimal_library_design(
  ddpca_outpath,
  position_offset = 0,
  fitness_abundance_file,
  fitness_binding_file,
  max_num_mutants,
  explore_mutations = "all_starting",
  RT = 0.001987 * (273 + 30),
  sample_size = 1000,
  optimal_perc = 0.7,
  min_num_doubles = 20,
  colour_scheme,
  outpath
)
}
\arguments{
\item{ddpca_outpath}{path to ddPCA data (required)}

\item{position_offset}{residue position offset (default:0)}

\item{fitness_abundance_file}{path to abundance fitness data (required)}

\item{fitness_binding_file}{path to binding fitness data (required)}

\item{max_num_mutants}{maximum number of mutations (required)}

\item{explore_mutations}{vector of mutation ids to explore or "all_starting" (default:"all_starting")}

\item{RT}{constant (default:0.001987*(273+30))}

\item{sample_size}{number of random samples (default:1000)}

\item{optimal_perc}{optimal percentage maximum geometric mean growth rate (default:0.7)}

\item{min_num_doubles}{minimum number of doubles per single for candidate mutations (default:20)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Library design.
}
