% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__pairwise_distances_from_PDB.R
\name{archstabms__pairwise_distances_from_PDB}
\alias{archstabms__pairwise_distances_from_PDB}
\title{archstabms__pairwise_distances_from_PDB}
\usage{
archstabms__pairwise_distances_from_PDB(
  input_file,
  chain_query = "A",
  chain_target = "B",
  min_dist = TRUE
)
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain_query}{query chain id (default:A)}

\item{chain_target}{target chain id (default:B)}

\item{min_dist}{minimum distance (default:TRUE)}
}
\value{
data.table with minimum inter-chain (side-chain) heavy atom distances or list of data.tables with distances (if min_dist==FALSE)
}
\description{
Calculate pairwise (side-chain) heavy atom distances or minimum thereof.
}
