% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__plot_ddPCAvalidation_scatter.R
\name{archstabms__plot_ddPCAvalidation_scatter}
\alias{archstabms__plot_ddPCAvalidation_scatter}
\title{archstabms__plot_ddPCAvalidation_scatter}
\usage{
archstabms__plot_ddPCAvalidation_scatter(
  input_dt,
  ddPCA_inpath,
  trait,
  report_outpath,
  highlight_colour = "red",
  RT = 0.001987 * (273 + 24)
)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{ddPCA_inpath}{path to ddPCA free energies (required)}

\item{trait}{trait name to plot: folding or binding (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{highlight_colour}{colour for highlights (default:red)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
Nothing
}
\description{
Plot correlation with validation data.
}
