% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__plot_fitness_replicates_cor.R
\name{archstabms__plot_fitness_replicates_cor}
\alias{archstabms__plot_fitness_replicates_cor}
\title{archstabms__plot_fitness_replicates_cor}
\usage{
archstabms__plot_fitness_replicates_cor(input_dt, output_file, colour_scheme)
}
\arguments{
\item{input_dt}{list of folder paths to fitness data (required)}

\item{output_file}{output file (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Plot fitness replicate correlations.
}
