% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__plot_model_performance.R
\name{archstabms__plot_model_performance}
\alias{archstabms__plot_model_performance}
\title{archstabms__plot_model_performance}
\usage{
archstabms__plot_model_performance(
  input_dt,
  report_outpath,
  trait,
  highlight_colour = "red"
)
}
\arguments{
\item{input_dt}{data.table with model free energy estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{trait}{trait name and suffix for output files (required)}

\item{highlight_colour}{colour for highlights (default:red)}
}
\value{
Nothing
}
\description{
Plot model performance.
}
