% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__plot_mutation_distributions.R
\name{archstabms__plot_mutation_distributions}
\alias{archstabms__plot_mutation_distributions}
\title{archstabms__plot_mutation_distributions}
\usage{
archstabms__plot_mutation_distributions(
  input_dt,
  report_outpath,
  dataset_name,
  colour_scheme
)
}
\arguments{
\item{input_dt}{data.table with fitness estimates (required)}

\item{report_outpath}{output path for scatterplots (required)}

\item{dataset_name}{dataset name and suffix for output files (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Plot mutation distributions.
}
