% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__predict_fitness_ddPCA.R
\name{archstabms__predict_fitness_ddPCA}
\alias{archstabms__predict_fitness_ddPCA}
\title{archstabms__predict_fitness_ddPCA}
\usage{
archstabms__predict_fitness_ddPCA(
  mochi_outpath,
  folding_energy,
  binding_energy,
  RT = 0.001987 * (273 + 24)
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{folding_energy}{folding energies for 3-state model (required)}

\item{binding_energy}{binding energies for 3-state model (required for binding fitness prediction)}

\item{RT}{constant (default:0.001987*(273+24))}
}
\value{
A list with folding and binding fitness predictions
}
\description{
Predict fitness from folding and binding energies for 3-state model.
}
