% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__pvalue.R
\name{archstabms__pvalue}
\alias{archstabms__pvalue}
\title{archstabms__pvalue}
\usage{
archstabms__pvalue(av, se, degreesFreedom = 5, mu = 0, testType = "ztest")
}
\arguments{
\item{av}{sample mean (required)}

\item{se}{sample standard error (required)}

\item{degreesFreedom}{an integer degrees of freedom (default:5)}

\item{mu}{population mean (default:0)}

\item{testType}{type of test: either "ztest" or "ttest" (default:"ztest")}
}
\value{
P-value
}
\description{
Obtain P-value from z test or t test
}
