% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__random_library_design.R
\name{archstabms__random_library_design}
\alias{archstabms__random_library_design}
\title{archstabms__random_library_design}
\usage{
archstabms__random_library_design(
  ddpca_outpath,
  position_offset = 0,
  fitness_abundance_file,
  fitness_binding_file,
  max_num_mutants,
  sample_size_order = 10000,
  load_mochi_data,
  colour_scheme,
  outpath
)
}
\arguments{
\item{ddpca_outpath}{path to ddPCA data (required)}

\item{position_offset}{residue position offset (default:0)}

\item{fitness_abundance_file}{path to abundance fitness data (required)}

\item{fitness_binding_file}{path to binding fitness data (required)}

\item{max_num_mutants}{maximum number of mutations (required)}

\item{sample_size_order}{number of samples per mutation order (default:1000)}

\item{load_mochi_data}{python (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Random library design.
}
