% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__reachable_single_AA_mutants.R
\name{archstabms__reachable_single_AA_mutants}
\alias{archstabms__reachable_single_AA_mutants}
\title{archstabms__reachable_single_AA_mutants}
\usage{
archstabms__reachable_single_AA_mutants(wt_coding_seq, position_offset = 0)
}
\arguments{
\item{wt_coding_seq}{WT coding sequence string (required)}

\item{position_offset}{residue position offset (default:0)}
}
\value{
List of reachable mutant IDs and a dictionary mapping IDs to codons
}
\description{
Determine single AA mutants reachable from single nucleotide substitutions.
}
