% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__relative_angle_from_PDB.R
\name{archstabms__relative_angle_from_PDB}
\alias{archstabms__relative_angle_from_PDB}
\title{archstabms__relative_angle_from_PDB}
\usage{
archstabms__relative_angle_from_PDB(input_file, chain = "A")
}
\arguments{
\item{input_file}{path to PDB file (required)}

\item{chain}{chain id (default:A)}
}
\value{
data.table with relative angles for all residues
}
\description{
Get relative angle of c-beta to c-alpha vector w.r.t. center of mass to c-alpha vector from PDB file.
}
