% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__saturation_library_design.R
\name{archstabms__saturation_library_design}
\alias{archstabms__saturation_library_design}
\title{archstabms__saturation_library_design}
\usage{
archstabms__saturation_library_design(
  pdb_file,
  pdb_chain_query = "A",
  pdb_chain_target = "B",
  colour_scheme,
  outpath
)
}
\arguments{
\item{pdb_file}{path to PDB file (required)}

\item{pdb_chain_query}{query chain id (default:A)}

\item{pdb_chain_target}{target chain id (default:B)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Saturation combinatorial mutagenesis library design.
}
