% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms__window_library_design.R
\name{archstabms__window_library_design}
\alias{archstabms__window_library_design}
\title{archstabms__window_library_design}
\usage{
archstabms__window_library_design(
  fitness_abundance_file,
  pdb_file,
  pdb_chain_query = "A",
  pdb_chain_target = "B",
  final_window_size = 22,
  final_window_pos = 10,
  colour_scheme,
  outpath
)
}
\arguments{
\item{fitness_abundance_file}{path to abundance fitness data (required)}

\item{pdb_file}{path to PDB file (required)}

\item{pdb_chain_query}{query chain id (default:A)}

\item{pdb_chain_target}{target chain id (default:B)}

\item{final_window_size}{final window size (default:22)}

\item{final_window_pos}{final window start position (default:10)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Saturation combinatorial mutagenesis library design.
}
