% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_binding_plots.R
\name{archstabms_binding_plots}
\alias{archstabms_binding_plots}
\title{archstabms_binding_plots}
\usage{
archstabms_binding_plots(
  dataset_name,
  coef_order,
  base_dir,
  colour_scheme,
  outpath
)
}
\arguments{
\item{dataset_name}{character dataset name (required)}

\item{coef_order}{coefficient order (required)}

\item{base_dir}{Base directory (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Plot fitness distributions and scatterplots.
}
