% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_coupling_decay.R
\name{archstabms_coupling_decay}
\alias{archstabms_coupling_decay}
\title{archstabms_coupling_decay}
\usage{
archstabms_coupling_decay(
  input_dir_list,
  min_mut_order,
  test_variants,
  outpath,
  colour_scheme
)
}
\arguments{
\item{input_dir_list}{list of MoCHI directories (required)}

\item{min_mut_order}{minimum mutation order (required)}

\item{test_variants}{whether to use unseen test variants only: T/F (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Coupling predictive decay plots.
}
