% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_coupling_heatmap.R
\name{archstabms_coupling_heatmap}
\alias{archstabms_coupling_heatmap}
\title{archstabms_coupling_heatmap}
\usage{
archstabms_coupling_heatmap(input_file, trait = "Folding", outpath)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{trait}{trait name (default:"Folding")}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Coupling scatterplots.
}
