% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_coupling_scatter.R
\name{archstabms_coupling_scatter}
\alias{archstabms_coupling_scatter}
\title{archstabms_coupling_scatter}
\usage{
archstabms_coupling_scatter(input_file, outpath, colour_scheme)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}
}
\value{
Nothing
}
\description{
Coupling scatterplots.
}
