% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_ddPCA_biophysical_model.R
\name{archstabms_ddPCA_biophysical_model}
\alias{archstabms_ddPCA_biophysical_model}
\title{archstabms_ddPCA_biophysical_model}
\usage{
archstabms_ddPCA_biophysical_model(
  mochi_outpath,
  temperature = 30,
  ddpca_outpath,
  position_offset,
  load_mochi_data,
  outpath
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{temperature}{temperature in degrees celcuis (default:30)}

\item{ddpca_outpath}{path to ddPCA data (required)}

\item{position_offset}{residue position offset (required)}

\item{load_mochi_data}{python (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Biophysical model from shallow ddPCA data.
}
