% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_ddPCA_linear_model_datasets.R
\name{archstabms_ddPCA_linear_model_datasets}
\alias{archstabms_ddPCA_linear_model_datasets}
\title{archstabms_ddPCA_linear_model_datasets}
\usage{
archstabms_ddPCA_linear_model_datasets(
  dataset_names,
  seq_position_offset,
  base_dir,
  output_dir,
  stagenum,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{seq_position_offset}{list of sequence position offsets (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Biophysical model from shallow ddPCA data for multiple datasets.
}
