% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_fitness_plots.R
\name{archstabms_fitness_plots}
\alias{archstabms_fitness_plots}
\title{archstabms_fitness_plots}
\usage{
archstabms_fitness_plots(dataset_name, base_dir, colour_scheme, outpath)
}
\arguments{
\item{dataset_name}{character dataset name (required)}

\item{base_dir}{Base directory (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{outpath}{output path for plots and saved objects (required)}
}
\value{
Nothing
}
\description{
Plot fitness distributions and scatterplots.
}
