% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_fitness_plots_datasets.R
\name{archstabms_fitness_plots_datasets}
\alias{archstabms_fitness_plots_datasets}
\title{archstabms_fitness_plots_datasets}
\usage{
archstabms_fitness_plots_datasets(
  dataset_names,
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Coupling scatterplots for multiple datasets.
}
