% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_library_design.R
\name{archstabms_library_design}
\alias{archstabms_library_design}
\title{archstabms_library_design}
\usage{
archstabms_library_design(
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Library design.
}
