% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_structure_metrics.R
\name{archstabms_structure_metrics}
\alias{archstabms_structure_metrics}
\title{archstabms_structure_metrics}
\usage{
archstabms_structure_metrics(
  input_file,
  outpath,
  pdb_file,
  getcontacts_WT_file,
  getcontacts_counts_file,
  pdb_chain_query = "A",
  pdb_chain_target = "B",
  execute = TRUE
)
}
\arguments{
\item{input_file}{path to MoCHI thermo model fit results (required)}

\item{outpath}{output path for plots and saved objects (required)}

\item{pdb_file}{path to PDB file (required)}

\item{getcontacts_WT_file}{path to getcontacts WT file (required)}

\item{getcontacts_counts_file}{path to getcontacts counts file (required)}

\item{pdb_chain_query}{query chain id (default:A)}

\item{pdb_chain_target}{target chain id (default:B)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Add 3D structure metrics.
}
