% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_structure_metrics_datasets.R
\name{archstabms_structure_metrics_datasets}
\alias{archstabms_structure_metrics_datasets}
\title{archstabms_structure_metrics_datasets}
\usage{
archstabms_structure_metrics_datasets(
  dataset_names,
  pdb_file_list,
  base_dir,
  output_dir,
  stagenum,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{pdb_file_list}{list of paths to PDB files (required)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Add 3D structure metrics of multiple datasets.
}
