% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_thermo_model_results.R
\name{archstabms_thermo_model_results}
\alias{archstabms_thermo_model_results}
\title{archstabms_thermo_model_results}
\usage{
archstabms_thermo_model_results(
  mochi_outpath,
  temperature = 30,
  literature_free_energies = NA,
  ddPCA_free_energies = NA,
  position_offset = 0,
  position_dict = list(),
  outpath,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{mochi_outpath}{path to MoCHI thermo model fit results (required)}

\item{temperature}{temperature in degrees celcuis (default:30)}

\item{literature_free_energies}{path to literature free energies (default:NA)}

\item{ddPCA_free_energies}{path to ddPCA free energies (default:NA)}

\item{position_offset}{residue position offset (default:0)}

\item{position_dict}{residue position dictionary (default:{})}

\item{outpath}{output path for plots and saved objects (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Evaluate thermo model results.
}
