% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archstabms_thermo_model_results_datasets.R
\name{archstabms_thermo_model_results_datasets}
\alias{archstabms_thermo_model_results_datasets}
\title{archstabms_thermo_model_results_datasets}
\usage{
archstabms_thermo_model_results_datasets(
  dataset_names,
  literature_free_energies,
  ddPCA_free_energies,
  base_dir,
  output_dir,
  stagenum,
  colour_scheme,
  execute = TRUE
)
}
\arguments{
\item{dataset_names}{character vector of dataset names (required)}

\item{literature_free_energies}{path to literature free energies (required)}

\item{ddPCA_free_energies}{path to ddPCA free energies (default:NA)}

\item{base_dir}{Base directory (required)}

\item{output_dir}{Output directory (required)}

\item{stagenum}{stage number (required)}

\item{colour_scheme}{colour scheme file (required)}

\item{execute}{whether or not to execute the analysis (default: TRUE)}
}
\value{
Nothing
}
\description{
Evaluate thermo model results of multiple datasets.
}
